requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
Pillow==10.0.0
argparse
tqdm==4.65.0
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
